move utils functions to separate files
- common
- render
- task
