+++ src/utils/common.js
@@ -1,0 +1,3 @@
+export const getRandomInteger = (min = 0, max = 1) => {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
